untitled file working copy - shared dispose handling
